Refactor.

+++ EasyDx/D3DHelpers.cpp
@@ -6,6 +6,6 @@ namespace dx
 {
     void SetName(ID3D11DeviceChild& child, std::string_view name)
     {
-        child.SetPrivateData(WKPDID_D3DDebugObjectName, name.size(), name.data());
+        child.SetPrivateData(WKPDID_D3DDebugObjectName, gsl::narrow<std::uint32_t>(name.size()), name.data());
     }
 }
